Issue #24: GitChange should be moved to data package
